Docker: run on non-x64, use ../maiko
run-medley assumes that maiko/ will be organized neatly with the
executables in the correct (platform-specific) folder, but we don't know
what platform we're on when we're building via Docker (and I don't think
that we should need to). To deal with this, we just add the entire maiko
folder to the PATH rather than only the subdirectory corresponding to
the current platform.

Build is still two-stage.