SEDIT-TOPLEVEL:  Add functions for setting/getting properties
also a field in the edit-context to store them, and a change to the region subroutine to respect the :REGION property.

Note that SEDIT has a peculiarity in that the record declarations are on SEDIT-DECLS but the runtime INITRECORDS is separately on SEDIT-ACCESS.  So you have to make sure you dump SEDIT-ACCESS, not just SEDIT-DECLS if a record changes.  A bit of a trap