TEDIT-FILE: Split up TEDIT.BUILD.PCTB into separate subfunctions
Should be easier to update and extend.  Not clear that the code for making a string piece is still needed here, since we interpret strings as file names.  So this may be further simplified.  I hope this also will give us a better handle on some of the file inconsistencies, by isolating the readers for formatted and unformatted files.