Another try: fixed the menus, maybe fixed hardcopy
The OPENSTRINGSTREAM always seemed to be in the neightborhood of the hard hardcopy crash, since it was being used to insert page numbers.  I changed that, and also changed the TEDITSTRING entry point to eliminate the dependency on OPENSTRINGSTREAM.  Instead, it opens an empty text stream, and then does TEDIT.INSERT of the desired string.

But note:  I never identified the problem with hardcopy, so still tread lightly.