Added READBOM and WRITEBOM
Also filled out the 16-BE/LE formats on top of the on-demand code-translation that was already there.  These are independent additions, just putting them in now to speed up the merging so that Tedit can start to deal with the bom.

It would be good for a BOM expert to verify that I got it right.