Mainly scrolling and line updating when Tedit windows are partially offscreen
This is hopefully getting near the end of the rework of scrolling aimed at avoiding flicker with the fast repetitive scrolling that wheelscroll can produce.  The idea is to calculate what part of the screen bitmap can just be moved around, including the selection highlighting, so that only newly revealed lines need to be redisplayed.  The easier cases are when the window is totally on screen.  Next easiest, in previous versions, is when the bottom of the window is offscreen.  This aims at the hardest case, where the top of the window is not visible.

This also includes a fix to the misnamed \TEDIT.COPYINSERTFN that @nbriggs noticed, plus perhaps some other issues that have been reported that I now don't remember.

It does not address the problem that @pamoroso recently noticed, where type-in selection is getting screwed up (and another glitch where type-in overflow at the end of the window moves up too many lines).