Rmk16  remove tedit special stuff from  CLIPBOARD, move it to TEDIT-COMMAND. (#1676)
* CLIPBOARD:  Remove Tedit-specific code

* TEDIT-COMMAND:  Add interface to Clipboard

This also implements a different way of dealing with image objects in the selection.  Instead of causing an error (since an imageobject can't be printed to the clipboard stream), it tries to put out a useful representation of the object.  If the object has a preprint function, it applies that.  Otherwise, it prints out the name of the getfn.  Other objects can be explored.