Fix Medley builds on github actions - currently breaks due to error in creating Docker image for arm64. (#2039)
In github actions, remove build of Medley Docker for arm64 since the Dockerfile_medley build fails when loading tigervnc-standalone-server for arm64.  Arm64 docker image has never been used anyway.  It was meant for running online on arm64 servers, but we have never done that and have no plans to do so.  So just getting rid of arm64 docker image rather than trying to fix it.