If CADR of a TEDIT.ABBREVS entry is LITATOM that can CHARCODE.DECODE (i.e., it's on CHARACTERNAMES), use it. Clear the TEdit prompt window if \TEDIT.ABBREV.EXPAND finds an expansion in TEDIT.ABBREVS.
"=" is duplicated as an expansion KEY, so modified TEDIT.ABBREVS entry:
("=" "357,121")
to
("lra" "357,121"  ; U+21C6 Leftwards Arrow Over Rightwards Arrow ("lra" is "left-right-arrow")
                )
because "=" is also the KEY in entry:
("=" NONBREAKING-HYPHEN)
which was never seen as a possibility (the above was earlier in the ALIST).
(This change exposed the issue with LITATOM as CADR of entry.)

Cleaned up some comments, typos. Added a few more TEDIT.ABBREVS entries.